requirements.txt
aiogram==3.0.0
python-dotenv==1.0.0
asyncio==3.4.3
logging==0.4.9.6
